+++ BTDB.SourceGenerator.Test/MetadataTests.cs
@@ -306,6 +306,29 @@ public class MetadataTests : GeneratorTestsBase
     }
 
     [Fact]
+    public Task VerifyDerivedGrandChildFromGeneratedClassHasMetadata()
+    {
+        // language=cs
+        return VerifySourceGenerator("""
+            namespace TestNamespace;
+
+            [BTDB.Generate]
+            public class Child
+            {
+                public ulong Id { get; set; }
+            }
+
+            public class DerivedChild : Child
+            {
+            }
+
+            public class GrandChild : DerivedChild
+            {
+            }
+            """);
+    }
+
+    [Fact]
     public Task VerifyDerivedClassFromGenericBaseUsesClosedOwnerTypeForMetadata()
     {
         // language=cs